requirements.txt
uvicorn
fastapi
fastapi-pagination
pytest
pytest-asyncio
SQLalchemy
marshmallow-sqlalchemy
psycopg2-binary
alembic
python-jose
passlib
python-dotenv
colorama
cloudinary
python-multipart
redis[hiredis]
elasticsearch==8.15.0
openai
stripe
app
